+++ src/aks-preview/azext_aks_preview/_params.py
@@ -820,6 +820,11 @@ def load_arguments(self, _):
             is_preview=True,
         )
         c.argument(
+            "enable_acns",
+            action="store_true",
+            is_preview=True,
+        )
+        c.argument(
             "custom_ca_trust_certificates",
             options_list=["--custom-ca-trust-certificates", "--ca-certs"],
             is_preview=True,
@@ -1324,6 +1329,16 @@ def load_arguments(self, _):
         )
         c.argument(
             "disable_fqdn_policy",
+            action="store_true",
+            is_preview=True,
+        )
+        c.argument(
+            "enable_acns",
+            action="store_true",
+            is_preview=True,
+        )
+        c.argument(
+            "disable_acns",
             action="store_true",
             is_preview=True,
         )