+++ src/aks-preview/azext_aks_preview/custom.py
@@ -607,6 +607,7 @@ def aks_create(
     custom_ca_trust_certificates=None,
     enable_advanced_network_observability=None,
     enable_fqdn_policy=None,
+    enable_acns=None,
     # nodepool
     crg_id=None,
     message_of_the_day=None,
@@ -841,6 +842,8 @@ def aks_update(
     disable_advanced_network_observability=None,
     enable_fqdn_policy=None,
     disable_fqdn_policy=None,
+    enable_acns=None,
+    disable_acns=None,
     # metrics profile
     enable_cost_analysis=False,
     disable_cost_analysis=False,